+++ core/execution/market.go
@@ -271,7 +271,7 @@ type Market struct {
 	positionFactor        num.Decimal // 10^pdp
 	assetDP               uint32
 
-	settlementDataInMarket *num.Uint
+	settlementDataInMarket *num.Numeric
 	nextMTM                time.Time
 	mtmDelta               time.Duration
 }
@@ -739,6 +739,7 @@ func (m *Market) OnTick(ctx context.Context, t time.Time) bool {
 	timer.EngineTimeCounterAdd()
 
 	m.updateMarketValueProxy()
+	m.updateLiquidityScores()
 	m.updateLiquidityFee(ctx)
 	m.broker.Send(events.NewMarketTick(ctx, m.mkt.ID, t))
 	return m.closed
@@ -771,7 +772,7 @@ func (m *Market) blockEnd(ctx context.Context) {
 				Price:         mp.Clone(),
 				OriginalPrice: mcmp,
 			}
-			m.confirmMTM(ctx, dummy, nil)
+			m.confirmMTM(ctx, dummy, nil, false)
 		}
 
 		closedWithoutLP := []events.MarketPosition{}
@@ -1119,7 +1120,7 @@ func (m *Market) leaveAuction(ctx context.Context, now time.Time) {
 		ID:            m.idgen.NextID(),
 		Price:         cmp,
 		OriginalPrice: mcmp,
-	}, updatedOrders)
+	}, updatedOrders, false)
 	// set next MTM
 	m.nextMTM = m.timeService.GetTimeNow().Add(m.mtmDelta)
 
@@ -1732,7 +1733,7 @@ func (m *Market) handleConfirmation(ctx context.Context, conf *types.OrderConfir
 }
 
 func (m *Market) confirmMTM(
-	ctx context.Context, order *types.Order, orderUpdates []*types.Order,
+	ctx context.Context, order *types.Order, orderUpdates []*types.Order, skipMargin bool,
 ) {
 	// now let's get the transfers for MTM settlement
 	markPrice := m.getLastTradedPrice()
@@ -1781,10 +1782,12 @@ func (m *Market) confirmMTM(
 	// force check
 	m.checkForReferenceMoves(
 		ctx, orderUpdates, false)
-	// release excess margin for all positions
-	pos := m.position.Positions()
-	// we can safely ignore the error here
-	_ = m.recheckMargin(ctx, pos)
+	if !skipMargin {
+		// release excess margin for all positions
+		pos := m.position.Positions()
+		// we can safely ignore the error here
+		_ = m.recheckMargin(ctx, pos)
+	}
 	// release any excess if needed
 	// m.releaseExcessMargin(ctx, pos...)
 }
@@ -2065,7 +2068,7 @@ func (m *Market) resolveClosedOutParties(ctx context.Context, distressedMarginEv
 		Price:         mp,
 		OriginalPrice: mcmp,
 	}
-	m.confirmMTM(ctx, dummy, nil)
+	m.confirmMTM(ctx, dummy, nil, false)
 
 	// Only check margins if MTM was successful.
 	return orderUpdates, m.recheckMargin(ctx, evt)
@@ -3242,7 +3245,7 @@ func (m *Market) tradingTerminated(ctx context.Context, tt bool) {
 				Price:         mp,
 				OriginalPrice: mcmp,
 			}
-			m.confirmMTM(ctx, dummy, nil)
+			m.confirmMTM(ctx, dummy, nil, true)
 		}
 		m.mkt.State = types.MarketStateTradingTerminated
 		m.mkt.TradingMode = types.MarketTradingModeNoTrading
@@ -3270,7 +3273,7 @@ func (m *Market) tradingTerminated(ctx context.Context, tt bool) {
 	m.log.Debug("market must not terminated before its enactment time", logging.MarketID(m.GetID()))
 }
 
-func (m *Market) settlementData(ctx context.Context, settlementData *num.Uint) {
+func (m *Market) settlementData(ctx context.Context, settlementData *num.Numeric) {
 	m.mu.Lock()
 	defer m.mu.Unlock()
 
@@ -3297,8 +3300,10 @@ func (m *Market) settlementDataWithLock(ctx context.Context) {
 		}
 
 		// mark price should be updated here
-		m.lastTradedPrice = settlementDataInAsset.Clone()
-		m.markPrice = settlementDataInAsset.Clone()
+		if settlementDataInAsset != nil {
+			m.lastTradedPrice = settlementDataInAsset.Clone()
+			m.markPrice = settlementDataInAsset.Clone()
+		}
 
 		// send the market data with all updated stuff
 		m.broker.Send(events.NewMarketDataEvent(ctx, m.GetMarketData()))
@@ -3409,6 +3414,9 @@ func (m *Market) distributeLiquidityFees(ctx context.Context) error {
 	if len(shares) == 0 {
 		return nil
 	}
+
+	// get liquidity scores and reset for next period
+	shares = m.updateSharesWithLiquidityScores(shares)
 
 	feeTransfer := m.fee.BuildLiquidityFeeDistributionTransfer(shares, acc)
 	if feeTransfer == nil {